refactor(desktop): extract useReader (loadMessage/preview/open) from App.tsx (F5.11)
Opening a message into the reading pane moves into useReader: loadMessage
(the openIdRef-guarded fetch that restores cached AI panels, resets the
per-message subsystems, and marks read), previewMessage (j/k, no
mark-read), and openMessage (Enter/click). Reader + AI state stays in App
and arrives via deps; the landmine-heavy loader is a verbatim move.

App.tsx 1,465 -> 1,351; useReader 174. unit 61 / e2e 39 green (open ->
summarize/prompt -> switch -> return caching covers it), tsc + build clean.
